refactor(server-cloudflare): extract env types and split worker into app + entry point
Types flow downward from env.ts, factory imports from env.ts,
and worker.ts becomes a thin re-export entry point for Wrangler.
