+++ packages/server-cloudflare/src/factory.ts
@@ -1,4 +1,4 @@
 import { createFactory } from 'hono/factory';
-import type { AppEnv } from './worker';
+import type { AppEnv } from './env';
 
 export const factory = createFactory<AppEnv>();
